+++ example.py
@@ -3,5 +3,3 @@ import telekit
 TOKEN: str = telekit.utils.read_token(".env")
 
 telekit.example(TOKEN)
-
-